pyproject.toml
[project]
name = "strategy-msxvi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gpxpy>=1.6.2",
    "matplotlib>=3.10.0",
    "numpy>=2.3.5",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.0.1",
    "scipy>=1.16.3",
    "tqdm>=4.67.1",
]
